# ENH: optimize Collection non-affine transform to call transform once