@jklymak 2018-06-20
Timing: there is a small hit for this PR for non-affine transforms that are not too slow.  i.e. if I remove the `time.sleep`:

With time.sleep:        34 s before, 1.09 s after
Without time.sleep:  0.9 s before, 1.10 s after

I'd argue the small 20% hit just for non-affine transforms is worth it to make slow transforms work much faster.