@ImportanceOfBeingErnest 2018-06-20
Ähm.. no. 99.999% of all transforms being used in matplotlib are expected to be sufficiently quick. So you say you would rather let those standard cases become slower????