@jklymak 2018-06-20
This code path is *only* for non-affine transforms.  Affine transforms aren't affected...