@ImportanceOfBeingErnest 2018-06-20
The standard polar plot contains a non-affine transform. Does this proposed solution make the creation of contours on polar plots slower?