@jklymak 2018-06-21
That probably depends on the nature of the contours....