@jklymak 2018-06-21
Old: 168 ms
New: 168 ms 
(20 iterations)