@jklymak 2018-07-04
OK, closing this, because I don't think it'll work for general implimentations of `transform_path_non_affine`  Thanks @dopplershift  for catching that!